Introduced Interactive Environment, Further explained below
Created VirtualEnvironment.py to serve the purpose of containing the overall environment that interfaces with the neural network, While ArcadeManager.py handles the code for running the arcade simulation itself.

EnvironmentObjectManager.py Handles creation of all static objects in the simulation.

Finally Keymap.py is a simply enum that maps Key Names to Key Codes

Signed-off-by: Will Richards <jwilt9000@gmail.com>