Simulation is in a state in which training/testing with neural networks can start
Added functionality to AgentController.py including calculating positive and negative rewards as well as being able
to step the agent forward and take an action, the ability to reset the simulation has also been added.
Collisions are now being handled properly and observations can be made about the environment to then use in an agent's decision.

Inside ArcadeManager.py a step method has been created to replace on_update, this allows for a more controllable fixed update.
A simple reset method has been created as a cleaner way to reset the environment.

One small change to the creation of the goal has been made so that it is easier to move the placement station around.

Created basic testing ground inside VirtualEnvironment.py to test the step method and rewards.
NOTE: The line `sleep(env.physics_environment.step_length)` is there to act as a pause so the simulation isn't always being updated this will likely be removed but exists there to allow for the developer to see what is happening

Finally, any development in the way of different neural networks should be done in a different branch

Signed-off-by: Will Richards <jwilt9000@gmail.com>